requirements.txt
streamlit>=1.32.0
yfinance>=0.2.37
pandas>=2.0.0
pandas-ta>=0.3.14b0
plotly>=5.18.0
google-genai>=2.0.0
lxml>=5.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
